Don't use MatchDisplay component
